cc_bot/requirements.txt
discord.py
python-dotenv
easy-pil